+++ Tools/visualData/visualCheckDataLabel.py
@@ -25,8 +25,8 @@ for file in filesList:
     rawFilename = file
     gtFilename = os.path.join(labelDir, patientID + suffix)
 
-    rawImage = np.load(rawFilename).astype(np.float32)
-    gtImage = np.load(gtFilename).astype(np.float32)
+    rawImage = np.load(rawFilename).astype(float)
+    gtImage = np.load(gtFilename).astype(float)
 
     # get sliceIndices for gtImage.
     nonzeroIndex = np.nonzero(gtImage)
